Remove unused packages as part of the startup script.
GPU instances use an Ubuntu based host OS, and so include a lot
of installed packages that we neither need nor use. This commit
extends the startup script for GPU instances to remove (some)
of those unused packages.

This issue does not apply to non-GPU instances, as those use
Container Optimized OS for their host image, and it has already
been pruned down to only include the software needed to run
Docker containers.

To start with, this cleanup step just removes dnsmasq.