Allow specifying the network name for Datalab instances.
This change also adds a warning when creating an instance if the
network has firewall rules other than the '...-allow-ssh' rule that
is added by the tool.

That new warning is meant to prevent users from accidentally creating
an instance in a network with too many sources/types of inbound traffic
allowed.

This fixes #1535