Bundling polymer resources by default (#1712)
This PR adds a few dependency behavior elements that only surfaced when bundling the resources, and adds bundling as the default behavior of the build system. A new -d | --debug option to the web sources build script disables that to help with debugging.

Note that the sources are still not minified. For some reason, this breaks the code, arbitrarily removing some class definitions. We can look into that later.

This helps bring down the number of requests to first meaningful paint significantly (201 -> 24). Load time on my local machine went down from 1.5s to 1.2s, should be better when server over network.

* working on bundling, minifying still breaks DatalabFile

* add build argument

* cleanup