Make the fake metadata address configurable. (#1721)
This makes it possible to configure the GCE metadata fake to serve on a
different host/port combination; the primary goal is to make it possible to
use the metadata service locally without the related Docker network
configs. (Without this, server startup will fail attempting to bind to
`metadata.google.internal`.)

I also let the metadata server pick up `settings.configDir`, since it was
already being threaded in.