Parallel file operations, unblock file browser (#1739)
This PR unblocks the file browser operations while a file list fetch is taking place. It does this by:

Making opening file/notebook a sync operation.
Making the _fetchFileList call non-blocking, so that it can be called while loading the file list. It follows from this that multiple calls to this method can be made, which is a nice addition.
Not blocking all other file operations on the fetch call, and making them a different _busy flag to enable the progress bar.
Note this doesn't fix another bug with double clicking items, where the breadcrumbs can go out of sync with the file list.

* use BaseFileManager for common functionality

* Use BaseFileManager, unified getNotebookUrl method

* double clicking items shouldn't be an async operation

* Use FileManager interface

* allow multiple parallel calls to _fetchFileList

* comment

* rebase changes

* let the last fetch request unset _fetching flag

* allow parallel async file operations

* merge master

* pr comments