Populate template notebooks in a new tab (#1753)
This PR makes populating template-based notebooks take place in a new tab as soon as the user picks a location for the notebook. The template name and parameters are sent as querystring parameters, and the notebook is populated and saved in the newly created tab, while showing a toast that says "Creating notebook.." This toast then disappears and the page is redirected to open the new notebook.
This does not get blocked by popup blockers, while also keeping the focus on the newly created notebook.
The PR takes the opportunity to refactor the way templating works around resource path resolution. Now the resolution is a simple Utils call, instead of keeping around a reference of some Polymer element.

* create template on the notebook page

* show toast while notebook is being created

* make template-manager a datalab-app dependency

* cleanup