Extend the reconnection message with more details.
This builds on the change from #1759 and extends it so that the printed command
can be directly copied and pasted in order to reconnect.

This introduces the potential to give the user a command to run that won't actually
work (e.g. if their connection failed due to the specified port being unavailable).

To prevent that risk, I've changed the reconnection message so that we only print
it if the initial connection succeeded.

This closes #1519