Pin web-component-tester to version 6.3.x
Yesterday, a new version of web-component-tester (6.4.0) was released,
and after that release our selenium tests (which run inside of
web-component-tester) started hanging forever.

Pinning to the older version seems to fix this issue.

There is one subtle issue where polymer_cli was installing the latest
version of web-component-tester, so the tests still failed even if we
later installed the older version. We avoid that issue by installing
the older version of web-component-tester before polymer_cli.