Add filter to item-list (#1805)
This PR adds filtering capability to item list. The filter only works on the first column of the item list for now, which covers the common cases. We can expand on this later as needed.

In order to do this, I switched the way we get the selection index to use the clicked item's true order in the original list, as opposed to its order in the rendered (filtered) list.

This design isn't final, it's just a rough way to get filter functionality into the UI. It's a toggle that shows up next to the first column in the header, which on click shows an input box underneath the header. Items will change as the user types in that box, without needing to hit Enter. Filtering is case-insensitive for now to simplify the implementation and the UI.

* filter toggle and input styling, no functionality

* filter working, still need to fix selected indices

* fix getting selected index

* add filter tests

* pr comments

* move resetting the filter to the file browser

* remove no longer needed test