Support base paths without the trailing slash.
Prior to this change, if you set a base path to `/foo/bar/`, then
requests to `/foo/bar/` would resolve to requests to the root path,
but requests to `/foo/bar` would result in a 404 error.

This would happen even if you set the base path to `/foo/bar`, as
the server normalizes the base path to include leading and trailing
slashes (to avoid accidental collisions with things like
`/foo/barbutnotreally`).

This change fixes that issue by explicitly checking for the case
that the request path matches the base path without the trailing slash.