Add sort to item-list (#1812)
This PR adds sorting capability to the item-list element. It does this by adding a new Column class that the users of item-list use to specify each column's name and type. Currently only 'string' and 'date' types are supported, but this list can easily extend to numbers for example.

The UI adds a down/up arrow icon next to the currently sorted column, and turns columns into clickable buttons. Clicking the currently sorted column again reverses the sort order.

* add column type to item-list columns

* add sort functionality to item-list, need UI

* tests for sorting

* sort UI, clicking columns works, test failing

* bail out on sort if no columns have been defined yet

* cleanup

* fix firefox tests

* add UI sort tests

* re-enable firefox tests

* support Date|number|string, only format to string on display

* express explicit dependency on columns

* add number sorting test

* add missing test for checking icons

* use index-as

* fix test with DATE type, add test for equal values

* pr comments