Switch from using the --zones flag to the "zone:(...)" filter
This change updates the `datalab list` command to incorporate the
`--zones` argument into the generated filter that is passed in to
the `gcloud compute instances list` command instead of forwarding
the `--zones` argument directly.

This is being changed because gcloud recently deprecated the `--zones`
flag in their tool, so users are now getting confusing deprecation
warnings from gcloud when they run `datalab list` with a `--zones`
argument.