Don't create a swapfile if it is too large.
This change adds a check to the Datalab instance startup script where
the swapfile creation/usage is prevented if the resulting swapfile
would use more than 10% of the available disk.

This is the first step towards addressing #1839, but the full issue
won't be fixed until we also provide a way to turn off swap entirely
for users who do not want it.