Silence a deprecation warning from gcloud.
The `datalab create` command is using a regex format for listing
the cloud source repositories that will be unsupported by a future
version of gcloud.

That causes gcloud to print a warning message (that does not get
presented to the user) about the format changing.

This commit fixes that by switching to the new format.