Improve the default verbosity. (#1861)
The previous default verbosity level was 'error'. This produces more
concise output, but it caused a good deal of user confusion when
the tool needed to perform a lot of setup steps (like creating a
network), and the user had no indication of what was going on.

The effect of that previous behavior was that the `datalab create`
command would seemingly hang for several minutes without giving
the user any indication whether or not it was making progress.

To avoid that issue, this change modifies the default verbosity
level to provide the user a notification every time the tool
starts to perform a potentially long-running operation.

More specifically, under the new behavior, the tool will print
(by default) a message before it:

1. Creates the network
2. Creates the firewall rule
3. Creates the source repository
4. Creates the notebooks persistent disk

Each of these messages is only printed if the setup step is actually
going to be performed. For instance, if the network already exists,
then no message is printed about the network.

Users who do not want this level of detail can override it by
setting the verbosity level to one of 'warning', 'error', 'critical',
or 'none'.