Check existing dialogs before opening new ones (#1862)
This fixes an issue where dialogs can be opened multiple times if the toolbar buttons are clicked too quickly.

Currently, we have no use case for opening multiple dialogs on top of each other, but this could change. This PR prevents multiple dialogs of the same type from opening, allowing an error dialog for example to appear on top of a new notebook dialog if we need.