Proxy requests by hostname in addition to URL path.
This change adds support for receiving requests to "${PORT}-dot-$(hostname)",
and forwarding those requests to another server in the container running
on that port.

This is similar to the existing support for proxying requests to
"/_port/${PORT}/", but with one important difference; since the port
number is taken from the host instead of the path, the requests do
not need to be modified and the other server does not need to know
about our proxying logic.

With this change, if I start Datalab locally with the following
command:

    docker run --rm -it \
        --hostname "$(hostname)" \
        -p 127.0.0.1:8082:8080 datalab

then I can send requests to the ungit server at port 8083 by running:

    curl -v 8083-dot-$(hostname):8082

(Assuming I have an /etc/hosts entry for "8083-dot-$(hostname)")