Fix a bug where the repo cloning was being interrupted.
This fixes an issue that appears to have only occurred with the
COS-based GPU instances, where the instance would fail to clone
the 'datalab-notebooks' repository.

The root cause was that the instance would reboot while the startup
script was running. That would interrupt it after the disk was
formatted, but before the repository was cloned. This, in turn
caused the repo to never be cloned, as subsequent startups would
see that the disk was already formatted.

This change tries to mitigate the likelihood of that happening
by moving the long running operation (downloading the Datalab
image to run the 'git clone' command) to before the disk is
formatted.