Fix build break (#1942)
For some reason, npm decided it didn't want to install one of the nested dependencies: . This is a direct dependency of polyserve, which is a dependency of polymer-cli, which in turn is a dependency of Datalab. This reproduces locally and on travis, and I'm not sure why this only happens with the latest polymer-cli (1.6.0).

I've also updated a few other dependencies, and put back the good old package-lock.json into git so it can help make these investigations easier in the future. (There's no need to review this file.)

* add @types/assert to fix build break, update deps

* add package-lock file

* put back ^'s