Add a facility to warn users of versions with known issues.
This change was prompted by the recent issue we had with version
20180105, where an issue was discovered, but we did not have a
good means of notifying users about the issue.

With this change, the CLI will download a versions info file from
https://storage.googleapis.com/cloud-datalab/version-issues.js
that contains a JSON object reporting known issues for any
affected versions of the CLI. It will then check the current
version against that object and report any matching warnings.