Migrate Docker Container to Conda (#1923)
* Migrates the Docker container from pip to Conda
with the exception of a few packages that are not available or
up-to-date in the Conda repositories.

Also updates some Python packages and cleans-up the Dockerfile some.

Additionally, the nbconvert config file is updated to disable xsrf token
checking in Jupyter so that tests will pass. One of the updated libraries
seems to have made the tokens a requirement by default. The tokens are
sometimes present in the request headers when running Datalab locally,
but the tokens are not being sent in tests for whatever reason. This
will be tracked via an issue and handled in a follow-up.

* Increasing notebook test timeout in attempt to stop remote failures. local runs pass consistently.

* Fixes conda environment / jupyter kernel setup. Now when a user switches kernels within the notebook, the conda environment is likewise switched. Additionally, updates the GPU container to use conda.

* Ensures new jupyter command in content/run.sh is called from within the Python3 conda environment. Given this, it's likely that jupyter should exist outside of the conda environments to prevent breakages like this in the future