Pin the version of the protobuf library installed.
Previously, we were picking the version of protobuf implicitly,
and the version in the Python 2 env (3.2.0) was different from
the version in the Python 3 env (3.5.2).

With the upgrade of the tensorflow library to 1.8.0, it is no longer
compatible with version 3.2.0 of the protobuf library, so this
is required for the new version of tensorflow to work in the
python 2 environment.