Track the known version issues in the source control history.
This change adds the 'version-issues.js' contents to the source
repository, and configures our cloud builder config to update
it in Google Cloud Storage on every push to the master branch.

The initial list of issues checked in includes the currently served
one with the addition of the issue about GPU instances not working
with the older NVIDIA driver.