Make the end-to-end CLI test use a free port for connections.
This change extends the end-to-end test for the CLI so that instead
of always using port 8081 (the default), it picks a port that was
free at the time the connection was being created.

This is meant to fix a recent regression in the tests where it would
fail in our Cloud Builder runs due to the port not being available.

There is still a race condition where the port could be grabbed by
another process between the time that it is picked and the time that
the connection is created, but that is not worth worrying about.