Pick a random zone for CLI tests.
This changes the CLI end-to-end test so that it defaults to
picking a random zone in the region 'us-west' instead of being
hard-coded to 'us-west1-a'.

The test will still honor the 'zone' config setting in gcloud if
it is set; this only changes the behavior when that is not set.