Rollback the upgrade of the GPU base image.
We previously switched from the cuda-9.0 base image to the
cuda-9.1 base image because the newer NVIDIA drivers seemed
to rely on cuda-9.1

However, Tensorflow is compiled against cuda-9.0, and cannot
even be imported with the cuda-9.1 image.

Conversely, using the cuda-9.0 image with the newest NVIDIA
drivers show that they still work with that older version
of the CUDA libraries.

As such, this commit rolls back that previous change and
puts as back at using the cuda-9.0 base image for GPU instances.