Update the version of Node JS installed when building.
This updates the portion of the build script that installs
NodeJS if it is not already installed on the build machine.

This mainly affects our Cloud Build config, as the build takes
place inside of a container that does not already have NodeJS
installed.

The reason for this change is that Polymer CLI recently updated,
and the new version fails if the version of NodeJS is older than
7.0. This change has broken our Cloud Build config, and is
blocking the Datalab release.