Fix a potential source of CI test flakiness.
Our automated Container-Builder test runs have been failing a lot
recenty with errors saying that the selected port for a Datalab
connection was already in use.

Since we pick a free port for the connection, and nothing else
is running in the container, the only explanation I can think of
is that the port has not yet been freed from when we initially
connected to it (in order to find a free port).

To fix this, we add a sleep of 10 seconds between picking the
port and trying to use it.