Retry integration tests on any exception.
The previous version of this logic retried a test if any of its
`self.assert...` calls failed.

However, that does not handle the case where a nested call to
`datalab` fails. In those cases a `CalledProcessError` would be
raised.

To retry on those scenarios as well, we extend the retry logic
to retry on any instance of the `Exception` class.