Fix an incompatibility introduced by new PyPI releases.
The latest version of `apache-beam` is incompatible with the latest
version of `dill` (both released within the past two weeks).

We were not previously pinning the versions of those packages, so the
result is that in newly built images you cannot import `apache_beam`.

That shows up in our ML Toolbox sample notebooks, as they indirectly
import apache_beam.

This change fixes that by pinning the versions of both `apache-beam`
and `dill`.