Remove the `apache-beam` library.
This was conflicting with the (older) `google-cloud-dataflow` library
which is what the `mltoolbox` modules in the `pydatalab` library are
written against.

We could alternatively update the mltoolbox modules to work with the
newer apache-beam library, but that is a more involved process.

The `apache-beam` installation was just added this week (when I was
trying to fix a regression caused by an update to `dill`), so no one
should be depending on this yet, as it was never included in a
released image.