Timeout Datalab connections if they never become healthy
This change adds a timeout to the health check, after which the
SSH tunnel will be killed (triggering it to be recreated), if
the health check never succeeds before that timeout.

This adds an important check to catch issues where the SSH tunnel
itself is unhealthy as opposed to the connecting instance being
unhealthy.

This fixes #2051