Add a test for internal-ip-only instances.
This change adds a test to the existing end-to-end test suite that
exercises the new functionality for creating an instance with no
external IP address.

Since the created instance is not accessible outside of the virtual
network, this test uses a bastion VM to create the SSH tunnel to that
instance.

Additionally, instances without external IP addresses cannot clone our
sample notebooks from GitHub, so we do not perform the check to make
sure the samples are available.